requirements.txt
python-telegram-bot==20.7
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
alembic==1.13.1
python-dotenv==1.0.0
pillow==10.1.0
aiofiles==23.2.1
apscheduler==3.10.4
cryptography==41.0.7
